New score function and better matcher layout
