+++ score/raw.go
@@ -1,0 +1,6 @@
+package score
+
+// Raw returns the distance as the score
+func Raw(query, target string, dist float64) float64 {
+	return dist
+}
